+++ rpcs3/Emu/RSX/Common/TextureUtils.h
@@ -5,6 +5,27 @@
 #include <vector>
 #include "Utilities/GSL.h"
 
+namespace rsx
+{
+	enum texture_upload_context
+	{
+		shader_read = 0,
+		blit_engine_src = 1,
+		blit_engine_dst = 2,
+		framebuffer_storage = 3
+	};
+
+	//Sampled image descriptor
+	struct sampled_image_descriptor_base
+	{
+		texture_upload_context upload_context = texture_upload_context::shader_read;
+		rsx::texture_dimension_extended image_type = texture_dimension_extended::texture_dimension_2d;
+		bool is_depth_texture = false;
+		f32 scale_x = 1.f;
+		f32 scale_y = 1.f;
+	};
+}
+
 struct rsx_subresource_layout
 {
 	gsl::span<const gsl::byte> data;
